redact resource_tools and file_metadata unused fns
