+++ src/crn_utils/_resource_tools.py
